+++ .gitignore
@@ -38,3 +38,6 @@ next-env.d.ts
 
 # cursor
 .cursorrules
+
+# turbo
+.turbo
